# docs/openapi-camareros.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Respuesta de error con código estable y mensaje en español.",
        "properties": {
          "code": {
            "examples": [
              "identity.credential_revoked"
            ],
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            ],
            "examples": [
              "Clave revocada. Renueva la clave"
            ],
            "title": "Detail"
          }
        },
        "required": [
          "code",
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SupresionRequest": {
        "description": "Confirmación (password) para ejercer el derecho de supresión.",
        "properties": {
          "password": {
            "examples": [
              "contraseña-mín-8"
            ],
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "title": "SupresionRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Identidad profesional (camareros) y credenciales/QR: ver AGENTS.md.",
    "title": "Personal Hostelería — Identity (profesionales)",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/camareros/ficha": {
      "get": {
        "description": "Ficha pública por QR `phid1` (sin token): solo campos visibles.",
        "operationId": "ficha_publica_v1_camareros_ficha_get",
        "parameters": [
          {
            "in": "query",
            "name": "qr",
            "required": true,
            "schema": {
              "title": "Qr",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CamareroFichaPublica"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Cuerpo de la petición inválido."
          }
        },
        "summary": "Ficha Publica",
        "tags": [
          "camareros"
        ]
      }
    },
    "/v1/camareros/ficha/foto": {
      "get": {
        "description": "Foto pública por QR (sin token): solo si `foto=true` y existe.",
        "operationId": "ficha_foto_v1_camareros_ficha_foto_get",
        "parameters": [
          {
            "in": "query",
            "name": "qr",
            "required": true,
            "schema": {
              "title": "Qr",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "image/webp": {}
            },
            "description": "Imagen de perfil (WebP)."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "El camarero no tiene foto de perfil."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Cuerpo de la petición inválido."
          }
        },
        "summary": "Ficha Foto",
        "tags": [
          "camareros"
        ]
      }
    },
    "/v1/camareros/ficha/foto/{camarero_id}": {
      "get": {
        "description": "Foto pública por id (para el directorio): solo si `foto=true` y existe.",
        "operationId": "ficha_foto_por_id_v1_camareros_ficha_foto__camarero_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "camarero_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Camarero Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "image/webp": {}
            },
            "description": "Imagen de perfil (WebP)."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "El camarero no tiene foto de perfil."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ficha Foto Por Id",
        "tags": [
          "camareros"
        ]
      }
    },
    "/v1/camareros/me": {
      "delete": {
        "operationId": "suprimir_cuenta_v1_camareros_me_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupresionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupresionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token de sesión inválido o caducado."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Cuerpo de la petición inválido."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Suprimir Cuenta",
        "tags": [
          "camareros"
        ]
      },
      "get": {
        "operationId": "me_v1_camareros_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CamareroPerfil"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token de sesión inválido o caducado."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Me",
        "tags": [
          "camareros"
        ]
      },
      "patch": {
        "operationId": "actualizar_me_v1_camareros_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PerfilUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CamareroPerfil"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token de sesión inválido o caducado."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Actualizar Me",
        "tags": [
          "camareros"
        ]
      }
    },
    "/v1/camareros/me/establecimientos": {
      "get": {
        "operationId": "mis_establecimientos_v1_camareros_me_establecimientos_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EstablecimientoMembresiaResponse"
                  },
                  "title": "Response Mis Establecimientos V1 Camareros Me Establecimientos Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token de sesión inválido o caducado."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mis Establecimientos",
        "tags": [
          "camareros"
        ]
      }
    },
...
    "/v1/camareros/me/invitaciones": {
      "get": {
        "operationId": "mis_invitaciones_v1_camareros_me_invitaciones_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InvitacionCamareroResponse"
                  },
                  "title": "Response Mis Invitaciones V1 Camareros Me Invitaciones Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token de sesión inválido o caducado."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Mis Invitaciones",
        "tags": [
          "camareros"
        ]
      }
    },
...
    "/v1/camareros/me/pagina-publica": {
      "put": {
        "description": "Opt-in del camarero para aparecer en la web pública de los\nestablecimientos donde trabaja (matriz AND con `mostrar_equipo`).",
        "operationId": "actualizar_pagina_publica_v1_camareros_me_pagina_publica_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaginaPublicaUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CamareroPerfil"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token de sesión inválido o caducado."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Actualizar Pagina Publica",
        "tags": [
          "camareros"
        ]
      }
    },
...
  }
}